+++ auth-server/src/config/mod.rs
@@ -2,3 +2,4 @@ pub mod tokenizer;
 pub mod session_manager;
 pub mod locale;
 pub mod ipify;
+pub mod sms_manager;